+++ install/setup.sh
@@ -28,7 +28,7 @@ An error occurred that prevented Mycodo from being installed!
 
 Please contact the developer by submitting a bug report at
 https://github.com/kizniche/Mycodo/issues with the pertinent
-excerpts from the setup log located at Mycodo/install/setup.log
+excerpts from the setup log located at ${INSTALL_DIRECTORY}/install/setup.log
 '
     echo "An error occurred. Exiting..." >&2
     exit 1
@@ -88,7 +88,7 @@ trap : 0
 IP=$(ip addr | grep 'state UP' -A2 | tail -n1 | awk '{print $2}' | cut -f1  -d'/')
 
 if [[ -z ${IP} ]]; then
-  IP="your.IP.here"
+  IP="your.IP.address.here"
 fi
 
 date
@@ -97,6 +97,6 @@ echo >&2 "
 ** Mycodo successfully installed without errors! **
 ***************************************************
 
-Go to https://${IP}/, or whatever your Pi's
+Go to https://${IP}/, or whatever your Raspberry Pi's
 IP address is, to create an admin user and log in.
 "